refactor(store): hide the run record codec
Keep run schema and locked transitions at the harness boundary while limiting the durable codec to crate internals.
