+++ crates/rimz/src/store/AGENTS.md
@@ -4,7 +4,7 @@ Local contract for `crates/rimz/src/store/` — durable workspace state. Extends
 
 ## Writes and durability
 
-- Every mutator and its public intent/outcome vocabulary lands under [`writer.rs`](./writer.rs): the façade owns the ordinary commit and log-boundary primitives, and `writer/` owns the debounce, lifecycle, publish, reset, queue, and reap branches. Reads on the `Store` handle stay lock-free; snapshot schema is consumed through its owning module.
+- Every store mutator and its public intent/outcome vocabulary lands under [`writer.rs`](./writer.rs): the façade owns the ordinary commit and log-boundary primitives, and `writer/` owns the debounce, lifecycle, publish, reset, queue, and reap branches. Supervised-run schema, transitions, and their workspace lock belong to `harness::run`; private `run_store` supplies only the durable codec. Store reset is the deliberate cross-record exception because it cancels active runs before rotating room state. Reads on the `Store` handle stay lock-free; snapshot schema is consumed through its owning module.
 - `workspace.lock` is the only cross-process serialization. Every writer is a short-lived CLI process, so a change that needs ordering takes the lock rather than introducing an in-process actor.
 - Every durable write routes through [`atomic.rs`](./atomic.rs), and every fsync syscall lives inside it; `cargo xtask invariants` rejects a `sync_all`/`sync_data` call anywhere else. No module hand-rolls its own atomic dance, and the event-log frame encoding stays beside its decoder in [`event_log/frame.rs`](./event_log/frame.rs).
 - [`writer/queue.rs`](./writer/queue.rs) owns all three message surfaces together: the live `messages.jsonl`, terminal text in the sibling `history.jsonl`, and terminal outcomes in the event log. Archive retention stays with event-log rotation, global orphan-temp collection with GC, and cache-local temp-sibling cleanup with [`atomic.rs`](./atomic.rs).
